backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-multipart>=0.0.6
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
PyYAML>=6.0.1
gunicorn>=21.2.0
python-dotenv>=1.0.0
